requirements.txt
groq>=0.13.0
rich>=13.7.0
pydantic>=2.6.0
python-dotenv>=1.0.0
